minor edits
I deleted "(bottles)" that was written each time after "batch" or "batches". I know that batch can also have another meaning than a recipient such as a bottle but I think it is quite clear what is meant here.
We should perhaps somewhere refer to a description how TS and VS have to be determined correctly, standard methods for example. However these are not for free I think.